pyproject.toml
[tool.poetry]
name = "gryffen"
version = "0.3.3"
description = "A FastAPI app that automates trading strategies with different brokers."
authors = [
    "Thomas Lin <ithomaslin@gmail.com>",
    "Thomas Lin <thomas@neat.tw>"
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = { version = "^0.23.2", extras = ["standard"] }
pydantic = {extras = ["email"], version = "^2.4.2"}
yarl = "^1.8.2"
ujson = "^5.7.0"
SQLAlchemy = {version = "^2.0.0", extras = ["asyncio"]}
alembic = "^1.9.2"
aiomysql = "^0.2.0"
aiofiles = "^23.2.1"
httptools = "^0.5.0"
loguru = "^0.6.0"
pymysql = "^1.0.3"
requests = "^2.28.2"
finnhub-python = "^2.4.17"
python-dotenv = "^1.0.0"
python-jose = "^3.3.0"
websocket-client = "^1.5.1"
python-multipart = "^0.0.6"
pyjwt = "^2.6.0"
aioredis = "^2.0.1"
schema = "^0.7.5"
google-auth = "^2.23.2"
google-api-python-client = "^2.87.0"
google-auth-oauthlib = "^1.0.0"
deprecated = "^1.2.14"
google = "^3.0.0"
mangum = "^0.17.0"
pre-commit = "^3.4.0"
pytest-asyncio = "^0.21.1"
jinja2 = "^3.1.2"
pydantic-settings = "^2.0.3"
google-cloud-storage = "^2.11.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
httpx = "^0.23.3"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["gryffen",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "GRYFFEN_ENVIRONMENT=pytest",
    "TEST_EMAIL=ithomaslin@gmail.com",
]

[fastapi-template.options]
project_name = "gryffen"
api_type = "rest"
enable_redis = "None"
enable_rmq = "None"
ci_type = "github"
enable_migrations = "True"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "gryffen"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "True"
traefik_labels = "None"
add_dummy = "None"
orm = "sqlalchemy"
self_hosted_swagger = "True"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
